requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-ckeditor==6.4.2
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.13.1
importlib-metadata==4.12.0
Markdown==3.4
Pillow==9.2.0
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
zipp==3.8.1
